Hoàn Thành 17 bai

+++ BaiTapThucTap/BaiTapThucTap/Controllers/BaiTap3Controller.cs
@@ -21,18 +21,18 @@ namespace BaiTapThucTap.Controllers
         }
         public IActionResult Index()
         {
-            var listSanPham = _db.tbl_DM_San_Pham.Include(x => x.LoaiSP).Include(x => x.DonViTinh).ToList();
+            var listSanPham = _db.tbl_DM_San_Pham.Include(x => x.LoaiSP).Include(x => x.DonViTinh).Where(x => x.Id != 1).ToList();
             return View(listSanPham);
         }
         private void LoadViewBag()
         {
-            ViewBag.LoaiSPList = _db.tbl_DM_Loai_San_Pham.Select(x => new SelectListItem
+            ViewBag.LoaiSPList = _db.tbl_DM_Loai_San_Pham.Where(x => x.Id != 1).Select(x => new SelectListItem
             {
                 Value = x.Id.ToString(),
                 Text = x.Ten_LSP
             }).ToList();
 
-            ViewBag.DonViTinhList = _db.tbl_DM_Don_Vi_Tinh.Select(x => new SelectListItem
+            ViewBag.DonViTinhList = _db.tbl_DM_Don_Vi_Tinh.Where(x => x.Id != 1).Select(x => new SelectListItem
             {
                 Value = x.Id.ToString(),
                 Text = x.Ten_Don_Vi_Tinh
@@ -76,17 +76,27 @@ namespace BaiTapThucTap.Controllers
                     ModelState.AddModelError("Ten_San_Pham", "Tên  sản phẩm đã tồn tại.");
                 }
             }
-            LoadViewBag();
+
             if (danhmuc.Loai_San_Pham_ID <= 0)
             {
                 ModelState.AddModelError("Loai_San_Pham_ID", "Vui lòng chọn loại sản phẩm.");
             }
-
             if (danhmuc.Don_Vi_Tin_ID <= 0)
             {
                 ModelState.AddModelError("Don_Vi_Tin_ID", "Vui lòng chọn đơn vị tính.");
             }
+            var listLoaiSanPham = _db.tbl_DM_Loai_San_Pham.ToList();
 
+            if (listLoaiSanPham.Find(x=>x.Id == danhmuc.Loai_San_Pham_ID) == null)
+            {
+                ModelState.AddModelError("Loai_San_Pham_ID", "Sản phẩm đã bị xóa");
+            }
+            LoadViewBag();
+            var listdonvitin = _db.tbl_DM_Don_Vi_Tinh.ToList();
+            if (listdonvitin.Find(x => x.Id == danhmuc.Don_Vi_Tin_ID) == null)
+            {
+                ModelState.AddModelError("Don_Vi_Tin_ID", "Đơn vị tính đã xóa");
+            }
             if (ModelState.IsValid)
             {
                 _db.tbl_DM_San_Pham.Add(danhmuc);
@@ -94,18 +104,28 @@ namespace BaiTapThucTap.Controllers
                 TempData["success"] = "Thêm Thành Công";
                 return RedirectToAction("Index");
             }
-
-            Console.WriteLine(ViewBag.DonViTinhList);
+            LoadViewBag();
             return View(danhmuc);
         }
 
-        public IActionResult Update(int id)
+        public IActionResult Update(int id,int idDVT,int idLSP)
         {
             var sp = _db.tbl_DM_San_Pham.Find(id);
+            var dvt  = _db.tbl_DM_Don_Vi_Tinh.Find(idDVT);
+            var lsp = _db.tbl_DM_Loai_San_Pham.Find(idLSP);
             if (sp == null)
             {
                 return NotFound();
             }
+            var viewModel = new BaiTap3
+            {
+                Ma_San_Pham=sp.Ma_San_Pham,
+                Ten_San_Pham = sp.Ten_San_Pham,
+                DonViTinh = dvt,
+                LoaiSP = lsp,
+                Ghi_Chu=sp.Ghi_Chu
+            };
+            LoadViewBag();
             return View(sp);
         }
         [HttpPost]
@@ -169,10 +189,27 @@ namespace BaiTapThucTap.Controllers
             {
                 ModelState.AddModelError("Don_Vi_Tin_ID", "Vui lòng chọn đơn vị tính.");
             }
+            var listLoaiSanPham = _db.tbl_DM_Loai_San_Pham.ToList();
 
-
-            if (ModelState.IsValid)
+            if (listLoaiSanPham.Find(x => x.Id == sp.Loai_San_Pham_ID) == null)
             {
+                ModelState.AddModelError("Loai_San_Pham_ID", "Sản phẩm đã bị xóa");
+            }
+            var listdonvitin = _db.tbl_DM_Loai_San_Pham.ToList();
+            if (listdonvitin.Find(x => x.Id == sp.Don_Vi_Tin_ID) == null)
+            {
+                ModelState.AddModelError("Don_Vi_Tin_ID", "Đơn vị tính đã xóa");
+            }
+            // Cập nhật các sản phẩm có khóa ngoại Don_Vi_Tin_ID trỏ đến đơn vị tính này
+            var sanPhamList7 = _db.tbl_DM_Nhap_Kho_Raw_Data.Where(x => x.San_Pham_ID == sp.Id).Count();
+            // Cập nhật các sản phẩm có khóa ngoại Don_Vi_Tin_ID trỏ đến đơn vị tính này
+            var sanPhamList11 = _db.tbl_DM_Xuat_Kho_Raw_Data.Where(x => x.San_Pham_ID == sp.Id).Count();
+            if (sanPhamList7 > 0 || sanPhamList11 > 0)
+            {
+                TempData["Error"] = "Sửa thất bại do sản phẩm đã tồn tại trong kho";
+                return RedirectToAction("Update", new { id = sp.Id });
+            }
+            if (ModelState.IsValid) {
                 existingSP.Ten_San_Pham = sp.Ten_San_Pham;
                 existingSP.Ma_San_Pham = sp.Ma_San_Pham;
                 existingSP.Ten_San_Pham = sp.Ten_San_Pham;
@@ -184,6 +221,7 @@ namespace BaiTapThucTap.Controllers
                 TempData["success"] = "Sửa Thành Công";
                 return RedirectToAction("Index");
             }
+                
             return View(sp);
         }
         //Hiển thị form xác nhận xóa chủng loại
@@ -196,14 +234,25 @@ namespace BaiTapThucTap.Controllers
             }
             return View(lsp);
         }
+        [HttpPost]
         // Xử lý xóa chủng loại
-        public IActionResult DeleteConfirmed(int id)
+        public IActionResult Delete(BaiTap3 model, int id)
         {
             var lsp = _db.tbl_DM_San_Pham.Find(id);
             if (lsp == null)
             {
                 return NotFound();
             }
+            // Cập nhật các sản phẩm có khóa ngoại Don_Vi_Tin_ID trỏ đến đơn vị tính này
+            var sanPhamList7 = _db.tbl_DM_Nhap_Kho_Raw_Data.Where(sp => sp.San_Pham_ID == id).Count();
+            // Cập nhật các sản phẩm có khóa ngoại Don_Vi_Tin_ID trỏ đến đơn vị tính này
+            var sanPhamList11 = _db.tbl_DM_Xuat_Kho_Raw_Data.Where(sp => sp.San_Pham_ID == id).Count();
+            if(sanPhamList7 >0 || sanPhamList11 >0)
+            {
+                TempData["Error"] = "Xóa thất bại do sản phẩm đã tồn tại trong kho";
+                return RedirectToAction("Delete", new {id=id });
+            }
+            
             _db.tbl_DM_San_Pham.Remove(lsp);
             _db.SaveChanges();
             TempData["success"] = "Xóa Thành Công";
